pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "resume-matcher"
version = "0.2.0"
description = "LLM-powered resume optimizer: parse, ATS-transform, score, tailor, cover letters, and Q&A"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "python-multipart>=0.0.9",
    "python-docx>=1.1.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "sqlalchemy>=2.0.0",
    "pdfplumber>=0.10.0",
    "pydantic-settings>=2.0.0",
    "google-genai>=1.0.0",
]

[tool.setuptools.packages.find]
include = ["resume_matcher*"]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "httpx>=0.27.0",
]
